Moved unit test parameter converters from Collection, Sequence, List, Set and Map and MutableMap into a separate package.
