+++ src/test/java/com/tsyba/core/collections/IndexRangeTests.java
@@ -1,6 +1,8 @@
 package com.tsyba.core.collections;
 
+import com.tsyba.core.collections.converter.IntList;
 import com.tsyba.core.collections.converter.IntOptional;
+import com.tsyba.core.collections.converter.StringList;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
 import org.junit.jupiter.params.converter.ArgumentConversionException;
