+++ src/test/java/com/tsyba/core/collections/ListTests.java
@@ -1,14 +1,12 @@
 package com.tsyba.core.collections;
 
+import com.tsyba.core.collections.converter.IntList;
 import com.tsyba.core.collections.converter.StringArray;
+import com.tsyba.core.collections.converter.StringCollection;
+import com.tsyba.core.collections.converter.StringList;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
-import org.junit.jupiter.params.converter.ArgumentConversionException;
-import org.junit.jupiter.params.converter.ConvertWith;
-import org.junit.jupiter.params.converter.TypedArgumentConverter;
 
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
 import java.util.Arrays;
 
 import static java.lang.String.format;
@@ -607,54 +605,6 @@ public class ListTests {
 			final var string = items.toString();
 			assertEquals(expected, string,
 				format("%s.toString()", items));
-		}
-	}
-}
-
-@Retention(RetentionPolicy.RUNTIME)
-@ConvertWith(StringList.Converter.class)
-@interface StringList {
-	@SuppressWarnings("rawtypes")
-	class Converter extends TypedArgumentConverter<String, List> {
-		protected Converter() {
-			super(String.class, List.class);
-		}
-
-		@Override
-		protected List<String> convert(String s) throws ArgumentConversionException {
-			if (s == null) {
-				return null;
-			} else {
-				final var items = new StringArray.Converter()
-					.convert(s);
-
-				return new List<>(items);
-			}
-		}
-	}
-}
-
-@Retention(RetentionPolicy.RUNTIME)
-@ConvertWith(IntList.Converter.class)
-@interface IntList {
-	@SuppressWarnings("rawtypes")
-	class Converter extends TypedArgumentConverter<String, List> {
-		protected Converter() {
-			super(String.class, List.class);
-		}
-
-		@Override
-		protected List<Integer> convert(String s) throws ArgumentConversionException {
-			if (s == null) {
-				return null;
-			} else {
-				final var converter = new StringArray.Converter();
-				final var items2 = Arrays.stream(converter.convert(s))
-					.map(Integer::parseInt)
-					.toArray(Integer[]::new);
-
-				return new List<>(items2);
-			}
 		}
 	}
 }
