+++ src/test/java/com/tsyba/core/collections/MapTests.java
@@ -1,18 +1,10 @@
 package com.tsyba.core.collections;
 
-import com.tsyba.core.collections.converter.StringArray;
-import com.tsyba.core.collections.converter.StringOptional;
+import com.tsyba.core.collections.converter.*;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
-import org.junit.jupiter.params.converter.ArgumentConversionException;
-import org.junit.jupiter.params.converter.ConvertWith;
-import org.junit.jupiter.params.converter.TypedArgumentConverter;
 
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
 import java.util.Arrays;
-import java.util.HashMap;
-import java.util.Objects;
 import java.util.Optional;
 
 import static java.lang.String.format;
@@ -783,81 +775,5 @@ class MapTests {
 		final var string = entries.toString();
 		assertEquals(expected, string,
 			format("%s.toString()", entries));
-	}
-}
-
-@Retention(RetentionPolicy.RUNTIME)
-@ConvertWith(StringMap.Converter.class)
-@interface StringMap {
-	@SuppressWarnings("rawtypes")
-	class Converter extends TypedArgumentConverter<String, Map> {
-		protected Converter() {
-			super(String.class, Map.class);
-		}
-
-		@Override
-		protected Map<String, String> convert(String s) throws ArgumentConversionException {
-			final var entries = new StringEntryArray.Converter()
-				.convert(s);
-
-			return new Map<>(entries);
-		}
-	}
-}
-
-@Retention(RetentionPolicy.RUNTIME)
-@ConvertWith(StringEntryArray.Converter.class)
-@interface StringEntryArray {
-	@SuppressWarnings("rawtypes")
-	class Converter extends TypedArgumentConverter<String, Map.Entry[]> {
-		protected Converter() {
-			super(String.class, Map.Entry[].class);
-		}
-
-		@Override
-		@SuppressWarnings("unchecked")
-		protected Map.Entry<String, String>[] convert(String s) throws ArgumentConversionException {
-			return Arrays.stream(new StringArray.Converter()
-					.convert(s))
-				.filter(Objects::nonNull)
-				.map(this::parseEntry)
-				.toArray(Map.Entry[]::new);
-		}
-
-		private Map.Entry<String, String> parseEntry(String s) {
-			final var parts = s.split("\\s*:\\s*");
-			final var key = "null".equals(parts[0])
-				? null
-				: parts[0];
-			final var value = "null".equals(parts[1])
-				? null
-				: parts[1];
-
-			return new Map.Entry<>(key, value);
-		}
-	}
-}
-
-@Retention(RetentionPolicy.RUNTIME)
-@ConvertWith(StringJavaMap.Converter.class)
-@interface StringJavaMap {
-	@SuppressWarnings("rawtypes")
-	class Converter extends TypedArgumentConverter<String, java.util.Map> {
-		protected Converter() {
-			super(String.class, java.util.Map.class);
-		}
-
-		@Override
-		protected java.util.Map<String, String> convert(String s) throws ArgumentConversionException {
-			final var entries = new StringEntryArray.Converter()
-				.convert(s);
-
-			final var map = new HashMap<String, String>();
-			for (var entry : entries) {
-				map.put(entry.key, entry.value);
-			}
-
-			return map;
-		}
 	}
 }
