+++ src/test/java/com/tsyba/core/collections/SequenceTests.java
@@ -1,20 +1,13 @@
 package com.tsyba.core.collections;
 
 import com.tsyba.core.collections.converter.IntOptional;
-import com.tsyba.core.collections.converter.StringArray;
+import com.tsyba.core.collections.converter.StringMap;
 import com.tsyba.core.collections.converter.StringOptional;
+import com.tsyba.core.collections.converter.StringSequence;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Nested;
-import org.junit.jupiter.params.converter.ArgumentConversionException;
-import org.junit.jupiter.params.converter.ConvertWith;
-import org.junit.jupiter.params.converter.TypedArgumentConverter;
 
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.util.Iterator;
 import java.util.Optional;
-import java.util.function.Function;
-import java.util.function.Predicate;
 
 import static java.lang.String.format;
 import static org.junit.jupiter.api.Assertions.assertEquals;
@@ -393,75 +386,6 @@ public class SequenceTests {
 				format("%s.enumerate(BiPredicate<Integer, T>)", items));
 			assertEquals(expected2, enumerated,
 				format("%s.enumerate(BiPredicate<Integer, T>)", items));
-		}
-	}
-}
-
-@Retention(RetentionPolicy.RUNTIME)
-@ConvertWith(StringSequence.Converter.class)
-@interface StringSequence {
-	@SuppressWarnings("rawtypes")
-	class Converter extends TypedArgumentConverter<String, Sequence> {
-		protected Converter() {
-			super(String.class, Sequence.class);
-		}
-
-		@Override
-		protected Sequence<String> convert(String s) throws ArgumentConversionException {
-			final var items = new StringArray.Converter()
-				.convert(s);
-
-			return new Sequence<>() {
-				@Override
-				public Sequence<String> getPrefix(int index) {
-					throw new UnsupportedOperationException();
-				}
-
-				@Override
-				public Sequence<String> getSuffix(int index) {
-					throw new UnsupportedOperationException();
-				}
-
-				@Override
-				public Sequence<String> get(IndexRange range) {
-					throw new UnsupportedOperationException();
-				}
-
-				@Override
-				public Sequence<String> matchAll(Predicate<String> condition) {
-					throw new UnsupportedOperationException();
-				}
-
-				@Override
-				public Sequence<Integer> findAll(String item) {
-					throw new UnsupportedOperationException();
-				}
-
-				@Override
-				public Sequence<Integer> findAll(Sequence<String> items) {
-					throw new UnsupportedOperationException();
-				}
-
-				@Override
-				public Collection<String> getDistinct() {
-					throw new UnsupportedOperationException();
-				}
-
-				@Override
-				public Sequence<String> reverse() {
-					throw new UnsupportedOperationException();
-				}
-
-				@Override
-				public <R> Sequence<R> convert(Function<String, R> converter) {
-					throw new UnsupportedOperationException();
-				}
-
-				@Override
-				public Iterator<String> iterator() {
-					return new ArrayIterator<>(items);
-				}
-			};
 		}
 	}
 }
